requirements.txt
requests
tkinter
Pillow
